requirements.txt
future
cffi<=1.12.3
pycparser==2.19
pymunk==5.5.0
numpy>=1.16.0
torch>=1.7 # check if you're on 3090 machine, may need different torch install
torchvision~=0.16.0
tensorboard>=2.0
matplotlib~=3.7.3
pillow~=10.1.0
h5py
six
progressbar2
opencv-python~=4.8.1.78
Cython~=3.0.3
marshmallow~=3.20.1
marshmallow_dataclass
PyYAML~=6.0.1
gym
yacs~=0.1.8
pygame
transforms3d
sympy
tinydb
jupyterlab
pandas~=2.0.3
imgaug
psutil
meshcat
pyglet<2
healpy
configargparse~=1.7
#open3d>=0.11
pynput>=1.7.3
ipython
PyMCubes
tqdm~=4.66.1
polyscope~=1.3.4
jedi==0.17.2  # to fix ipython bug
plotly>=4.14
kaleido>=0.2.1  # needed for plotly .png rendering
trimesh>=3.9
rtree>=0.9.7  # needed for trimesh contains method
networkx>=2.3
pybullet==3.1.3

scikit-learn~=1.3.2
scipy~=1.10.1
scikit-image~=0.21.0
pyvista~=0.42.3
open3d~=0.17.0
setuptools~=68.0.0
click~=8.1.7